// src/shared/types/generated/application/ItemDetails.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ItemId } from "./ItemId";
import type { ItemKind } from "./ItemKind";
import type { Lifecycle } from "./Lifecycle";
import type { MediaDetails } from "./MediaDetails";

export type ItemDetails = { item_id: ItemId, kind: ItemKind, lifecycle: Lifecycle, label: string | null, cover_media_item_id: ItemId | null, folder_ids: number[], media: Array<MediaDetails>, aggregate_tags: Array<string>, revision: number, };